Decouple Block Node simulators from SubProcessNetwork
