+++ hedera-node/test-clients/src/main/java/com/hedera/services/bdd/spec/utilops/BlockNodeSimulatorOp.java
@@ -3,7 +3,6 @@ package com.hedera.services.bdd.spec.utilops;
 
 import com.hedera.hapi.block.protoc.PublishStreamResponseCode;
 import com.hedera.services.bdd.junit.hedera.simulator.BlockNodeSimulatorController;
-import com.hedera.services.bdd.junit.hedera.subprocess.SubProcessNetwork;
 import com.hedera.services.bdd.spec.HapiSpec;
 import java.io.IOException;
 import java.util.concurrent.atomic.AtomicLong;
@@ -41,11 +40,8 @@ public class BlockNodeSimulatorOp extends UtilOp {
 
     @Override
     protected boolean submitOp(HapiSpec spec) throws Throwable {
-        if (!(spec.targetNetworkOrThrow() instanceof SubProcessNetwork network)) {
-            throw new IllegalStateException("Block node simulator operations require a SubProcessNetwork");
-        }
-
-        BlockNodeSimulatorController controller = network.getBlockNodeSimulatorController();
+        BlockNodeSimulatorController controller =
+                HapiSpec.TARGET_BLOCK_NODE_NETWORK.get().getBlockNodeSimulatorController();
         long verifiedBlock = 0;
 
         switch (action) {
